Move `ErrorKind` to separate module
